+++ packages/fern-dashboard/package.json
@@ -79,6 +79,7 @@
     "@tiptap/core": "3.0.5",
     "@tiptap/extension-drag-handle": "^3.0.9",
     "@tiptap/extension-drag-handle-react": "^3.0.9",
+    "@tiptap/extension-file-handler": "^3.4.1",
     "@tiptap/extension-placeholder": "^3.0.5",
     "@tiptap/extension-unique-id": "^3.0.9",
     "@tiptap/pm": "3.0.5",
